requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
